pyproject.toml
[project]
name = "crewai-agent"
version = "0.1.0"
description = "Multi-agent task planner and executor using crewai, pydantic, and async Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.76.0",
    "pydantic>=2.8.0",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
    "litellm[proxy]>=1.53.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.128.0",
]
